pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "codegate"
version = "0.1.0"
description = "Contract-driven evaluator for Python AI-generated code"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Jihad Hammoud"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]

[project.scripts]
codegate = "codegate.cli:main"

[tool.setuptools]
packages = ["codegate", "codegate.contract", "codegate.engine", "codegate.rules"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Enable coverage for subprocesses (used by the big end-to-end CLI test)
addopts = [
    "--cov=codegate",
    "--cov-report=term-missing",
    "--cov-report=xml",
]

[tool.coverage.run]
parallel = true
branch = true
source = ["codegate"]

[tool.coverage.report]
show_missing = true

[tool.coverage.paths]
source = [
    "codegate",
    "*/codegate",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
